+++ tests/e2e/__init__.py
@@ -1,0 +1,2 @@
+import os
+test_e2e_dir = os.path.dirname(os.path.realpath(__file__))