+++ tests/e2e/e2e_test_config.template.py
@@ -1,0 +1,3 @@
+# Copy this file to `e2e_test_config.py` and modified it, which is ignored by `.gitignore`
+handle='your handle'
+password='your password'