+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Changed
 
 - New project
+- use aiohttp
+- remove requirement.txt
+- extract http as class and make it injectable
+- add auto doc
+- add auto test(e2e, unit, coverage report)
 
 ### Features